feat: split MuqaConnectButton
